Move the rest of the VMs and adjust the references
