+++ 3DPoC/POC3D/POC3D.ViewModel/Calculations/ConstantMatrix.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Linq;
 
-namespace POC3D.Model.Calculations
+namespace POC3D.ViewModel.Calculations
 {
     public class ConstantMatrix
     {
