+++ 3DPoC/POC3D/POC3D.ViewModel/Commands/DeleteNodeCommand.cs
