+++ 3DPoC/POC3D/POC3D.ViewModel/ElementViewModel.cs
@@ -1,9 +1,8 @@
 ﻿using System.ComponentModel;
 using System.Windows.Media;
 using System.Windows.Media.Media3D;
-using POC3D.Helpers;
 using POC3D.Model;
-using POC3D.Model.Calculations;
+using POC3D.ViewModel.Calculations;
 using POC3D.ViewModel.Geometry;
 
 namespace POC3D.ViewModel
