+++ 3DPoC/POC3D/POC3D.ViewModel/NewElementViewModel.cs
