+++ 3DPoC/POC3D/POC3D.ViewModel/ProblemViewModel.cs
@@ -5,9 +5,8 @@ using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Input;
 using System.Windows.Media.Media3D;
-using POC3D.Helpers;
 using POC3D.Model;
-using POC3D.Model.Calculations;
+using POC3D.ViewModel.Calculations;
 using POC3D.ViewModel.Commands;
 
 namespace POC3D.ViewModel
