+++ 3DPoC/POC3D/POC3D.ViewModel/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="MathNet.Numerics" version="4.9.0" targetFramework="net48" />
   <package id="WindowsBase" version="4.6.1055.0" targetFramework="net48" />
 </packages>
